Explicit cast as required by NEC compiler.
Tom


git-svn-id: https://svn.cactuscode.org/flesh/trunk@1439 17b73243-c579-4c4c-a9d2-2d5706c11dac

+++ src/main/Groups.c
@@ -1558,7 +1558,7 @@ static CCTK_INT **CCTKi_ExtractSize(int dimension, const char *thorn, const char
           
           if(next_comma)
           {
-            strncpy(tmp, last_comma, (next_comma-last_comma));
+            strncpy(tmp, last_comma, (size_t)(next_comma-last_comma));
             tmp[next_comma-last_comma] = '\0';
           }
           else
